Converted activities to fragments
