+++ app/src/main/res/values/strings.xml
@@ -27,4 +27,7 @@
 
     <!-- Category name for the vocabulary words for family members [CHAR LIMIT=20] -->
     <string name="category_family">Family Members</string>
+
+    <!-- placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
